FIX: Update update_disk_usage to support php8.1 if filesize() returns an error [t33859]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@22630 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/resource_functions.php
@@ -6061,7 +6061,7 @@ function update_disk_usage($resource)
 		{
 		if ($f!=".." && $f!=".")
 			{
-			$s=filesize_unlimited($dir . "/" .$f);
+            $s=(int) filesize_unlimited($dir . "/" .$f);
 			$total+=$s;
 			}
 		}
